create user-form.component and use them in create-user and edit-user components - refactoring the scss for using the BEM parttner
