+++ angular14-firebase/src/app/user/components/create-user/create-user.component.html
@@ -1,100 +1,1 @@
-<div class="container">
-  <h2 class="title"> Create user</h2>
-  <form [formGroup]="userForm" (submit)="save()">
-
-    <mat-form-field appearance="outline">
-      <mat-label>Name</mat-label>
-      <input matInput formControlName="name" placeholder="Name">
-      <mat-icon matSuffix color="primary">person</mat-icon>
-      <mat-error *ngIf="userForm.get('name')?.hasError('required')">Name is required</mat-error>
-      <mat-error *ngIf="!userForm.get('name')?.hasError('required') && userForm.get('name')?.hasError('minlength')">
-        Name must be at least 2 characters long</mat-error>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Last name</mat-label>
-      <input matInput formControlName="lastName" placeholder="Last name">
-      <mat-icon matSuffix color="primary">person</mat-icon>
-      <mat-error *ngIf="userForm.get('lastName')?.hasError('required')">Last name is required</mat-error>
-      <mat-error
-        *ngIf="!userForm.get('lastName')?.hasError('required') && userForm.get('lastName')?.hasError('minlength')">
-        Last name must be at least 2 characters long</mat-error>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Email</mat-label>
-      <input matInput formControlName="email" placeholder="Email">
-      <mat-icon matSuffix color="primary">email</mat-icon>
-      <mat-error *ngIf="userForm.get('email')?.hasError('required')">Email is required</mat-error>
-      <mat-error *ngIf="!userForm.get('email')?.hasError('required') && userForm.get('email')?.hasError('email')">
-        Please provide a valid email</mat-error>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Phone Number</mat-label>
-      <input matInput mask="(00) 0 0000-0000" formControlName="phoneNumber" placeholder="Phone Number">
-      <mat-icon matSuffix color="primary">settings_cell</mat-icon>
-      <mat-error *ngIf="userForm.get('phoneNumber')?.hasError('required')">Phone number is required</mat-error>
-      <mat-error
-        *ngIf="!userForm.get('phoneNumber')?.hasError('required') && userForm.get('phoneNumber')?.hasError('pattern')">
-        Please provide a valid phone number</mat-error>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Salary expectation</mat-label>
-      <input matInput type="text" mask="separator.2" thousandSeparator="." formControlName="salaryExpectation"
-        placeholder="Salary expectation">
-      <mat-icon matSuffix color="primary">attach_money</mat-icon>
-      <mat-error *ngIf="userForm.get('salaryExpectation')?.hasError('required')">Salary expectation is required
-      </mat-error>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Level</mat-label>
-      <mat-select formControlName="level">
-        <mat-option value="{{developersLevelsEnum.JUNIOR}}">Junior</mat-option>
-        <mat-option value="{{developersLevelsEnum.MIDLEVEL}}">Mid-Level</mat-option>
-        <mat-option value="{{developersLevelsEnum.SENIOR}}">Senior</mat-option>
-        <mat-option value="{{developersLevelsEnum.SPECIALIST}}">Specialist</mat-option>
-        <mat-option value="{{developersLevelsEnum.TECHLEAD}}">Tech-Lead</mat-option>
-      </mat-select>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Principal stack</mat-label>
-      <mat-select formControlName="principalStack">
-        <mat-option value="{{principalStackEnum.FULLTSACK}}">Full Stack</mat-option>
-        <mat-option value="{{principalStackEnum.BACKEND}}">Back-End</mat-option>
-        <mat-option value="{{principalStackEnum.FRONTEND}}">Front-End</mat-option>
-        <mat-option value="{{principalStackEnum.MOBILE}}">Mobile</mat-option>
-        <mat-option value="{{principalStackEnum.DEVOPS}}">DevOps</mat-option>
-      </mat-select>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Technologies and Frameworks</mat-label>
-      <mat-chip-list #chipList aria-label="Tech selection">
-        <mat-chip *ngFor="let tech of techs" (removed)="remove(tech)">
-          {{tech}}
-          <button type="button" matChipRemove>
-            <mat-icon>cancel</mat-icon>
-          </button>
-        </mat-chip>
-        <input placeholder="New item..." #techInput [formControl]="techCtrl" [matAutocomplete]="auto"
-          [matChipInputFor]="chipList" [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-          (matChipInputTokenEnd)="add($event)">
-      </mat-chip-list>
-      <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
-        <ng-container *ngFor="let tech of filteredTechs | async">
-          <mat-option *ngIf="!techs.includes(tech)" [value]="tech">
-            {{tech}}
-          </mat-option>
-        </ng-container>
-      </mat-autocomplete>
-    </mat-form-field>
-
-    <button class="btn" type="submit" [class.spinner]="loading"
-      [ngClass]="userForm.invalid? '' : 'animate__animated animate__pulse animate__repeat-3'"
-      [disabled]="userForm.invalid || loading" mat-raised-button color="primary"> Create </button>
-  </form>
-</div>
+<app-user-form [title]="title"></app-user-form>
