--- angular14-firebase/src/app/user/components/edit-user/edit-user.component.scss
@@ -1,61 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.title {
-  margin-top: 2rem !important;
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-  margin-top: 1.5rem;
-  width: 65%;
-}
-
-.btn {
-  width: 250px;
-  height: 45px;
-  width: 100%;
-  margin-bottom: 1.5rem;
-}
-
-@media only screen and (max-width: 498px) {
-    mat-card {
-        height: 100%;
-        margin: unset;
-      }
-
-      .title {
-        margin-top: 7rem;
-      }
-
-      .container {
-        width: 100%;
-      }
-
-      form {
-        width: 100%;
-      }
-}
-
-@media only screen and (min-width: 498px) and (max-width: 960px) {
-    mat-card {
-        height: 100%;
-        margin: unset;
-      }
-
-      .title {
-        margin-top: 7rem;
-      }
-
-      .container {
-        width: 100%;
-      }
-
-      form {
-        width: 70%;
-      }
-}
