+++ angular14-firebase/src/app/user/components/user-form/user-form.component.scss
@@ -1,0 +1,51 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  &__title {
+    margin-top: 2rem !important;
+  }
+
+  &__form {
+    display: flex;
+    flex-direction: column;
+    margin-top: 1.5rem;
+    width: 65%;
+  }
+
+  &__btn {
+    width: 250px;
+    height: 45px;
+    width: 100%;
+    margin-bottom: 1.5rem;
+  }
+}
+
+@media only screen and (max-width: 498px) {
+  .container {
+    width: 100%;
+
+    &__title {
+      margin-top: 7rem;
+    }
+
+    &__form {
+      width: 100%;
+    }
+  }
+}
+
+@media only screen and (min-width: 498px) and (max-width: 960px) {
+  .container {
+    width: 100%;
+
+    &__title {
+      margin-top: 7rem;
+    }
+
+    &__form {
+      width: 70%;
+    }
+  }
+}
